django_api/requirements.txt
django==4.2.11
flake8==6.1.0
python-dotenv==1.0.1
django-split-settings==1.3.0
psycopg2-binary==2.9.9
icecream~=2.1.3